requirements.txt
awesome-slugify==1.6.5
certifi==2018.1.18
chardet==3.0.4
docker==3.0.0
docker-pycreds==0.2.1
idna==2.6
pymongo==3.6.0
regex==2018.1.10
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
websocket-client==0.46.0
